refactor: simplify hive.ts file into types, queries and hive
